move transfer_context.py into worker_transfer/ package
